[core][iOS] Separate the public API from the rest of the core
