+++ packages/expo-modules-core/ios/Api/Factories/ModuleFactories.swift
@@ -1,0 +1,6 @@
+/**
+ Sets the name of the module that is exported to the JavaScript world.
+ */
+public func Name(_ name: String) -> AnyDefinition {
+  return ModuleNameDefinition(name: name)
+}
