+++ packages/expo-modules-core/ios/Protocols/AnyDefinition.swift
@@ -1,0 +1,4 @@
+/**
+ A protocol that must be implemented to be a part of module's definition and the module definition itself.
+ */
+public protocol AnyDefinition {}
