--- packages/expo-modules-core/ios/Views/ViewFactory.swift
@@ -1,16 +1,0 @@
-/**
- A definition of the view factory that creates views.
- */
-internal struct ViewFactory: ViewManagerDefinitionComponent {
-  typealias FactoryClosureType = () -> UIView
-
-  let factory: FactoryClosureType
-
-  init(_ factory: @escaping FactoryClosureType) {
-    self.factory = factory
-  }
-
-  func create() -> UIView {
-    return factory()
-  }
-}
